+++ quotegen/src/main/java/com/andmark/quotegen/controller/api/BookController.java
@@ -34,4 +34,5 @@ public class BookController {
         ExtractedLinesDTO extractedLinesDTO = bookService.processPageAndLineNumber(requestDTO);
         return ResponseEntity.ok(extractedLinesDTO);
     }
+
 }