+++ quotegen/src/main/java/com/andmark/quotegen/service/WebVersionService.java
@@ -3,4 +3,6 @@ package com.andmark.quotegen.service;
 public interface WebVersionService {
     String getReadmeContent();
     String getChangelogContent();
+
+    String getPrivatePolicyContent();
 }